requirements.txt
flask==2.3.3
pydantic==2.5.2
python-dotenv==1.0.0
PyJWT==2.8.0
supabase==2.1.0
gunicorn==21.2.0 